// src/@types/src/openWeather.types.ts
export type WeatherSkyProps = {
  id: number;
  main: string;
  description: string;
  icon: string;
};

type OpenWeatherCurrentType = {
  dt: number;
  sunrise: number;
  sunset: number;
  temp: number;
  feels_like: number;
  pressure: number;
  humidity: number;
  dew_point: number;
  uvi: number;
  clouds: number;
  visibility: number;
  wind_speed: number;
  wind_deg: number;
  wind_gust: number;
  weather: Array<WeatherSkyProps>;
};

type OpenWeatherDailyType = {
  dt: number;
  sunrise: number;
  sunset: number;
  moonrise: number;
  moonset: number;
  moon_phase: number;
  temp: {
    day: number;
    min: number;
    max: number;
    night: number;
    eve: number;
    morn: number;
  };
  feels_like: {
    day: number;
    night: number;
    eve: number;
    morn: number;
  };
  pressure: number;
  humidity: number;
  dew_point: number;
  wind_speed: number;
  wind_deg: number;
  wind_gust: number;
  weather: Array<WeatherSkyProps>;
  clouds: number;
  pop: number;
  rain: number;
  uvi: number;
};

type OpenWeatherHoursType = {
  dt: number;
  temp: number;
  feels_like: number;
  pressure: number;
  humidity: number;
  dew_point: number;
  uvi: number;
  clouds: number;
  visibility: number;
  wind_speed: number;
  wind_deg: number;
  wind_gust: number;
  weather: Array<WeatherSkyProps>;
  pop: number;
};

export type OpenWeatherDataProps = {
  lat: number;
  lon: number;
  timezone: string;
  timezone_offset: number;
  current: OpenWeatherCurrentType;
  hourly: Array<OpenWeatherHoursType>;
  daily: Array<OpenWeatherDailyType>;
};